feat: use sdtlib path blob
On-behalf-of: SAP <matthias.bruns@sap.com>
